Emission + AlphaTexture + Albedo's Alpha visual tests
