+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Materials/Tests/PBRMaterialVisualTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections;
+using System.Linq;
 using DCL;
 using DCL.Components;
 using DCL.Helpers;
@@ -35,6 +36,114 @@ public class PBRMaterialVisualTests : VisualTestsBase
         }, camTarget, out IDCLEntity entity);
         yield return matPBR.routine;
 
+
+        yield return new WaitForAllMessagesProcessed();
+
+        yield return VisualTestHelpers.TakeSnapshot();
+    }
+
+    [UnityTest]
+    [VisualTest]
+    [Explicit]
+    [Category("Explicit")]
+    public IEnumerator StandardConfigurations_Generate() { yield return VisualTestHelpers.GenerateBaselineForTest(StandardConfigurations()); }
+
+    [UnityTest]
+    [VisualTest]
+    public IEnumerator StandardConfigurations()
+    {
+        yield return InitVisualTestsScene("PBRMaterialVisualTests_StandardConfigurations");
+
+        Vector3 camTarget = new Vector3(8, 3, 8);
+        VisualTestHelpers.RepositionVisualTestsCamera(VisualTestController.i.camera, new Vector3(9, 4, 17), camTarget);
+
+        GLTFShape gltfShape = TestHelpers.CreateEntityWithGLTFShape(scene, Vector3.zero,
+            new LoadableShape.Model
+            {
+                src = TestAssetsUtils.GetPath() + "/GLB/MaterialsScene.glb"
+            }, out IDCLEntity entity);
+        TestHelpers.SetEntityTransform(scene, entity, new DCLTransform.Model { position = new Vector3(0, 0, 8), rotation = Quaternion.Euler(90, 180, 0) });
+
+        LoadWrapper loader = GLTFShape.GetLoaderForEntity(entity);
+        yield return new WaitUntil(() => loader.alreadyLoaded);
+
+        yield return VisualTestHelpers.TakeSnapshot();
+    }
+
+    [UnityTest]
+    [VisualTest]
+    [Explicit]
+    [Category("Explicit")]
+    public IEnumerator TransparentObjectsAndSSAO_Generate() { yield return VisualTestHelpers.GenerateBaselineForTest(TransparentObjectsAndSSAO()); }
+
+    [UnityTest]
+    [VisualTest]
+    [Explicit]
+    [Category("Explicit")] //Enable the test once we properly render opaque objects with SSAO behind transparents
+    public IEnumerator TransparentObjectsAndSSAO()
+    {
+        yield return InitVisualTestsScene("PBRMaterialVisualTests_TransparentObjectsAndSSAO");
+
+        VisualTestHelpers.SetSSAOActive(true);
+
+        Vector3 camTarget = new Vector3(5, 1, 5);
+        VisualTestHelpers.RepositionVisualTestsCamera(VisualTestController.i.camera, new Vector3(4.6f, 1.8f, 0.6f), camTarget);
+
+        PlaneShape plane = TestHelpers.CreateEntityWithPlaneShape(scene, new Vector3(5, 1, 5), true);
+        IDCLEntity planeEntity = plane.attachedEntities.FirstOrDefault();
+        TestHelpers.SetEntityTransform(scene, planeEntity, new Vector3(5, 1, 5), Quaternion.identity, Vector3.one * 3);
+        TestHelpers.AttachPBRMaterialToEntity(scene, planeEntity, new PBRMaterial.Model { alphaTest = 1, transparencyMode = 1, albedoColor = Vector4.one });
+
+        BoxShape box1 = TestHelpers.CreateEntityWithBoxShape(scene, new Vector3(4, 1, 6), true);
+        TestHelpers.AttachPBRMaterialToEntity(scene, box1.attachedEntities.FirstOrDefault(), new PBRMaterial.Model { transparencyMode = 0, albedoColor = Color.blue });
+
+        BoxShape box2 = TestHelpers.CreateEntityWithBoxShape(scene, new Vector3(5, 1, 6.5f), true);
+        TestHelpers.AttachPBRMaterialToEntity(scene, box2.attachedEntities.FirstOrDefault(), new PBRMaterial.Model { transparencyMode = 0, albedoColor = Color.red });
+        yield return new WaitForAllMessagesProcessed();
+
+        yield return VisualTestHelpers.TakeSnapshot();
+    }
+
+    [UnityTest]
+    [VisualTest]
+    [Explicit]
+    [Category("Explicit")]
+    [TestCase(1, 100, ExpectedResult = null)]
+    [TestCase(0.2f, 100, ExpectedResult = null)]
+    [TestCase(1f, 10, ExpectedResult = null)]
+    public IEnumerator Emission_AlphaTexture_AlbedoAlpha_Generate(float alpha, float emissiveIntensity) { yield return VisualTestHelpers.GenerateBaselineForTest(Emission_AlphaTexture_AlbedoAlpha(alpha, emissiveIntensity)); }
+
+    [UnityTest]
+    [VisualTest]
+    [TestCase(1, 100, ExpectedResult = null)]
+    [TestCase(0.2f, 100, ExpectedResult = null)]
+    [TestCase(1f, 10, ExpectedResult = null)]
+    public IEnumerator Emission_AlphaTexture_AlbedoAlpha(float alpha, float emissiveIntensity)
+    {
+        yield return InitVisualTestsScene($"PBRMaterialVisualTests_Emission_AlphaTexture_AlbedoAlpha_{alpha}_{emissiveIntensity}");
+
+        VisualTestHelpers.SetSSAOActive(true);
+
+        Vector3 camTarget = new Vector3(5, 1, 5);
+        VisualTestHelpers.RepositionVisualTestsCamera(VisualTestController.i.camera, new Vector3(4.6f, 1.8f, 0.6f), camTarget);
+
+        DCLTexture texture = TestHelpers.CreateDCLTexture(scene, TestAssetsUtils.GetPath() + "/Images/avatar.png");
+        DCLTexture alphaTexture = TestHelpers.CreateDCLTexture(scene, TestAssetsUtils.GetPath() + "/Images/alphaTexture.png");
+        DCLTexture emissionTexture = TestHelpers.CreateDCLTexture(scene, TestAssetsUtils.GetPath() + "/Images/Gradient A4.png");
+
+        PlaneShape plane = TestHelpers.CreateEntityWithPlaneShape(scene, new Vector3(5, 1, 5), true);
+        IDCLEntity planeEntity = plane.attachedEntities.FirstOrDefault();
+        TestHelpers.SetEntityTransform(scene, planeEntity, new Vector3(5, 1, 5), Quaternion.Euler(0, 0, 180), Vector3.one * 3);
+        TestHelpers.AttachPBRMaterialToEntity(scene, planeEntity, new PBRMaterial.Model
+        {
+            albedoColor = new Color(1, 1, 1, alpha),
+            transparencyMode = 2,
+            albedoTexture = texture.id,
+            alphaTexture = alphaTexture.id,
+            emissiveColor = Color.blue,
+            emissiveIntensity = emissiveIntensity,
+            emissiveTexture = emissionTexture.id
+        });
 
         yield return new WaitForAllMessagesProcessed();
