+++ unity-renderer/TestResources/VisualTests/BaselineImages/PBRMaterialVisualTests_Emission_AlphaTexture_AlbedoAlpha_0,2_100_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:782ee25556cc7a18a18b3a43f53681753cd6f4f10539014c91f46caf48cc113a
+size 385927
