+++ unity-renderer/TestResources/VisualTests/BaselineImages/PBRMaterialVisualTests_Emission_AlphaTexture_AlbedoAlpha_1_100_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:39a5eda19a31578b97f86f7339cedefd6724b5050056836b3156f9c8b47ee80c
+size 404684
