+++ unity-renderer/TestResources/VisualTests/BaselineImages/PBRMaterialVisualTests_Emission_AlphaTexture_AlbedoAlpha_1_10_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a956bc165868a92118c97b9320dc66114c72e60495194eaf0d214daa07064037
+size 388088
